Displaying session costs in the physician and the patient's side

+++ esteshari/database/migrations/2023_06_11_00071054_create_physician_pricings_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('physician_pricings', function (Blueprint $table) {
             $table->id();
-            $table->double('cost')->default('0.00');
-            $table->double('discountedCost')->nullable();
+            $table->decimal('cost', 10,2)->default('0.00');
+            $table->decimal('discountedCost', 10,2)->nullable();
             $table->string('currency')->default('MYR');
             $table->timestamps();
         });
